Move spell events and spell preparation form to creature (#3980)
