+++ src/module/actor/sheet/base.ts
@@ -2,13 +2,12 @@ import type { ActorPF2e, CharacterPF2e } from "@actor";
 import { ActorDataPF2e } from "@actor/data";
 import { RollFunction } from "@actor/data/base";
 import { SAVE_TYPES } from "@actor/values";
-import { ItemPF2e, PhysicalItemPF2e, SpellcastingEntryPF2e, SpellPF2e } from "@item";
+import { ItemPF2e, PhysicalItemPF2e } from "@item";
 import { createConsumableFromSpell } from "@item/consumable/spell-consumables";
 import { ItemSourcePF2e } from "@item/data";
 import { isPhysicalData } from "@item/data/helpers";
 import { Coins } from "@item/physical/data";
 import { DENOMINATIONS } from "@item/physical/values";
-import { SpellPreparationSheet } from "@item/spellcasting-entry/sheet";
 import { DropCanvasItemDataPF2e } from "@module/canvas/drop-canvas-data";
 import { RollOptionRuleElement } from "@module/rules/rule-element/roll-option";
 import { createSheetTags, maintainTagifyFocusInRender, processTagifyInSubmitData } from "@module/sheet/helpers";
@@ -36,7 +35,6 @@ import { AddCoinsPopup } from "./popups/add-coins-popup";
 import { IdentifyItemPopup } from "./popups/identify-popup";
 import { RemoveCoinsPopup } from "./popups/remove-coins-popup";
 import { ScrollWandPopup } from "./popups/scroll-wand-popup";
-import { createSpellcastingDialog } from "./spellcasting-dialog";
 import { ActorSizePF2e } from "../data/size";
 
 /**
@@ -320,17 +318,6 @@ export abstract class ActorSheetPF2e<TActor extends ActorPF2e> extends ActorShee
         // Spell Create
         $spellControls.filter(".spell-create").on("click", (event) => this.onClickCreateItem(event));
 
-        // Adding/Editing/Removing Spellcasting entries
-        $spellcasting
-            .find("[data-action=spellcasting-create]")
-            .on("click", (event) => this.createSpellcastingEntry(event));
-        $spellControls
-            .filter("a[data-action=spellcasting-edit]")
-            .on("click", (event) => this.editSpellcastingEntry(event));
-        $spellControls
-            .filter("a[data-action=spellcasting-remove]")
-            .on("click", (event) => this.removeSpellcastingEntry(event));
-
         /* -------------------------------------------- */
         /*  Inventory                                   */
         /* -------------------------------------------- */
@@ -438,30 +425,6 @@ export abstract class ActorSheetPF2e<TActor extends ActorPF2e> extends ActorShee
             ]);
         });
 
-        // Update max slots for Spell Items
-        $html.find(".prepared-toggle").on("click", async (event) => {
-            event.preventDefault();
-            const itemId = $(event.currentTarget).parents(".item-container").attr("data-container-id") ?? "";
-            this.openSpellPreparationSheet(itemId);
-        });
-
-        $html.find(".slotless-level-toggle").on("click", async (event) => {
-            event.preventDefault();
-
-            const itemId = $(event.currentTarget).parents(".item-container").attr("data-container-id") ?? "";
-            const itemToEdit = this.actor.items.get(itemId);
-            if (!itemToEdit?.isOfType("spellcastingEntry"))
-                throw new Error("Tried to toggle visibility of slotless levels on a non-spellcasting entry");
-            const bool = !(itemToEdit.system.showSlotlessLevels || {}).value;
-
-            await this.actor.updateEmbeddedDocuments("Item", [
-                {
-                    _id: itemId ?? "",
-                    "system.showSlotlessLevels.value": bool,
-                },
-            ]);
-        });
-
         // Select all text in an input field on focus
         $html.find<HTMLInputElement>("input[type=text], input[type=number]").on("focus", (event) => {
             event.currentTarget.select();
@@ -474,17 +437,6 @@ export abstract class ActorSheetPF2e<TActor extends ActorPF2e> extends ActorShee
             if (match) target.value = match[0];
             else target.value = "";
         });
-    }
-
-    /** Opens the spell preparation sheet, but only if its a prepared entry */
-    openSpellPreparationSheet(entryId: string) {
-        const entry = this.actor.items.get(entryId);
-        if (entry?.isOfType("spellcastingEntry") && entry.isPrepared) {
-            const $book = this.element.find(`.item-container[data-container-id="${entry.id}"] .prepared-toggle`);
-            const offset = $book.offset() ?? { left: 0, top: 0 };
-            const sheet = new SpellPreparationSheet(entry, { top: offset.top - 60, left: offset.left + 200 });
-            sheet.render(true);
-        }
     }
 
     async onClickDeleteItem(event: JQuery.TriggeredEvent): Promise<void> {
@@ -682,82 +634,10 @@ export abstract class ActorSheetPF2e<TActor extends ActorPF2e> extends ActorShee
 
     /** Handle a drop event for an existing Owned Item to sort that item */
     protected override async _onSortItem(event: ElementDragEvent, itemSource: ItemSourcePF2e): Promise<ItemPF2e[]> {
-        const $dropItemEl = $(event.target).closest(".item");
-        const $dropContainerEl = $(event.target).closest(".item-container");
-
-        const dropSlotType = $dropItemEl.attr("data-item-type");
-        const dropContainerType = $dropContainerEl.attr("data-container-type");
         const item = this.actor.items.get(itemSource._id);
         if (!item) return [];
 
-        // if they are dragging onto another spell, it's just sorting the spells
-        // or moving it from one spellcastingEntry to another
-        if (item.isOfType("spell")) {
-            const targetLocation = $dropContainerEl.attr("data-container-id") ?? "";
-            const collection = this.actor.spellcasting.collections.get(targetLocation, { strict: true });
-
-            if (dropSlotType === "spellLevel") {
-                const { level } = $dropItemEl.data();
-                const spell = await collection.addSpell(item, { slotLevel: Number(level) });
-                this.openSpellPreparationSheet(collection.id);
-                return [spell ?? []].flat();
-            } else if ($dropItemEl.attr("data-slot-id")) {
-                const dropId = Number($dropItemEl.attr("data-slot-id"));
-                const slotLevel = Number($dropItemEl.attr("data-slot-level"));
-
-                if (Number.isInteger(dropId) && Number.isInteger(slotLevel)) {
-                    const allocated = await collection.prepareSpell(item, slotLevel, dropId);
-                    if (allocated) return [allocated];
-                }
-            } else if (dropSlotType === "spell") {
-                const dropId = $dropItemEl.attr("data-item-id") ?? "";
-                const target = this.actor.items.get(dropId);
-                if (target?.isOfType("spell") && item.id !== dropId) {
-                    const sourceLocation = item.system.location.value;
-
-                    // Inner helper to test if two spells are siblings
-                    const testSibling = (item: SpellPF2e, test: SpellPF2e) => {
-                        if (item.isCantrip !== test.isCantrip) return false;
-                        if (item.isCantrip && test.isCantrip) return true;
-                        if (item.isFocusSpell && test.isFocusSpell) return true;
-                        if (item.level === test.level) return true;
-                        return false;
-                    };
-
-                    if (sourceLocation === targetLocation && testSibling(item, target)) {
-                        const siblings = collection.filter((spell) => testSibling(item, spell));
-                        await item.sortRelative({ target, siblings });
-                        return [target];
-                    } else {
-                        const spell = await collection.addSpell(item, { slotLevel: target.level });
-                        this.openSpellPreparationSheet(collection.id);
-                        return [spell ?? []].flat();
-                    }
-                }
-            } else if (dropContainerType === "spellcastingEntry") {
-                // if the drop container target is a spellcastingEntry then check if the item is a spell and if so update its location.
-                // if the dragged item is a spell and is from the same actor
-                if (CONFIG.debug.hooks)
-                    console.debug("PF2e System | ***** spell from same actor dropped on a spellcasting entry *****");
-
-                const dropId = $(event.target).parents(".item-container").attr("data-container-id");
-                return dropId ? [await item.update({ "system.location.value": dropId })] : [];
-            }
-        } else if (item.isOfType("spellcastingEntry")) {
-            // target and source are spellcastingEntries and need to be sorted
-            if (dropContainerType === "spellcastingEntry") {
-                const sourceId = item.id;
-                const dropId = $dropContainerEl.attr("data-container-id") ?? "";
-                const source = this.actor.items.get(sourceId);
-                const target = this.actor.items.get(dropId);
-
-                if (source && target && source.id !== target.id) {
-                    const siblings = this.actor.spellcasting.contents;
-                    source.sortRelative({ target, siblings });
-                    return [target];
-                }
-            }
-        } else if (item.isOfType("physical")) {
+        if (item.isOfType("physical")) {
             const $target = $(event.target).closest("[data-item-id]");
             const targetId = $target.attr("data-item-id") ?? "";
             const target = this.actor.inventory.get(targetId);
@@ -797,7 +677,6 @@ export abstract class ActorSheetPF2e<TActor extends ActorPF2e> extends ActorShee
         return await this._onDropItem({ preventDefault(): void {} } as ElementDragEvent, data);
     }
 
-    /** Extend the base _onDropItem method to handle dragging spells onto spell slots. */
     protected override async _onDropItem(event: ElementDragEvent, data: DropCanvasItemDataPF2e): Promise<ItemPF2e[]> {
         event.preventDefault();
 
@@ -824,6 +703,21 @@ export abstract class ActorSheetPF2e<TActor extends ActorPF2e> extends ActorShee
             return [item];
         }
 
+        return this._handleDroppedItem(event, item, data);
+    }
+
+    /**
+     * PF2e specific method called by _onDropItem() when this is a new item that needs to be dropped into the actor
+     * that isn't already on the actor or transferring to another actor.
+     */
+    protected async _handleDroppedItem(
+        event: ElementDragEvent,
+        item: ItemPF2e,
+        data: DropCanvasItemDataPF2e
+    ): Promise<ItemPF2e[]> {
+        const actor = this.actor;
+        const itemSource = item.toObject();
+
         // mystify the item if the alt key was pressed
         if (event.altKey && item.isOfType("physical") && isPhysicalData(itemSource)) {
             itemSource.system.identification.unidentified = item.getMystifiedData("unidentified");
@@ -831,7 +725,6 @@ export abstract class ActorSheetPF2e<TActor extends ActorPF2e> extends ActorShee
         }
 
         // get the item type of the drop target
-        const $itemEl = $(event.target).closest(".item");
         const $containerEl = $(event.target).closest(".item-container");
         const containerAttribute = $containerEl.attr("data-container-type");
         const unspecificInventory = this._tabs[0]?.active === "inventory" && !containerAttribute;
@@ -841,13 +734,7 @@ export abstract class ActorSheetPF2e<TActor extends ActorPF2e> extends ActorShee
         // otherwise they are dragging a new spell onto their sheet.
         // we still need to put it in the correct spellcastingEntry
         if (item.isOfType("spell") && itemSource.type === "spell") {
-            if (dropContainerType === "spellcastingEntry") {
-                const entryId = $containerEl.attr("data-container-id") ?? "";
-                const collection = this.actor.spellcasting.collections.get(entryId, { strict: true });
-                const level = Math.max(Number($itemEl.attr("data-level")) || 0, item.baseLevel);
-                this.openSpellPreparationSheet(collection.id);
-                return [(await collection.addSpell(item, { slotLevel: level })) ?? []].flat();
-            } else if (dropContainerType === "actorInventory" && itemSource.system.level.value > 0) {
+            if (dropContainerType === "actorInventory" && itemSource.system.level.value > 0) {
                 const popup = new ScrollWandPopup(
                     this.actor,
                     {},
@@ -1040,66 +927,6 @@ export abstract class ActorSheetPF2e<TActor extends ActorPF2e> extends ActorShee
         }
 
         this.actor.createEmbeddedDocuments("Item", [data]);
-    }
-
-    /** Handle creating a new spellcasting entry for the actor */
-    private createSpellcastingEntry(event: JQuery.ClickEvent) {
-        event.preventDefault();
-        createSpellcastingDialog(event, this.actor);
-    }
-
-    private editSpellcastingEntry(event: JQuery.ClickEvent): void {
-        const { containerId } = $(event.target).closest("[data-container-id]").data();
-        const entry = this.actor.spellcasting.get(containerId, { strict: true });
-        createSpellcastingDialog(event, entry as Embedded<SpellcastingEntryPF2e>);
-    }
-
-    /**
-     * Handle removing an existing spellcasting entry for the actor
-     */
-    private removeSpellcastingEntry(event: JQuery.ClickEvent): void {
-        event.preventDefault();
-
-        const li = $(event.currentTarget).parents("[data-container-id]");
-        const itemId = li.attr("data-container-id") ?? "";
-        const item = this.actor.items.get(itemId);
-        if (!item) {
-            return;
-        }
-
-        // Render confirmation modal dialog
-        renderTemplate("systems/pf2e/templates/actors/delete-spellcasting-dialog.html").then((html) => {
-            new Dialog({
-                title: "Delete Confirmation",
-                content: html,
-                buttons: {
-                    Yes: {
-                        icon: '<i class="fa fa-check"></i>',
-                        label: "Yes",
-                        callback: async () => {
-                            console.debug("PF2e System | Deleting Spell Container: ", item.name);
-                            // Delete all child objects
-                            const itemsToDelete: string[] = [];
-                            for (const item of this.actor.itemTypes.spell) {
-                                if (item.system.location.value === itemId) {
-                                    itemsToDelete.push(item.id);
-                                }
-                            }
-                            // Delete item container
-                            itemsToDelete.push(item.id);
-                            await this.actor.deleteEmbeddedDocuments("Item", itemsToDelete);
-
-                            li.slideUp(200, () => this.render(false));
-                        },
-                    },
-                    cancel: {
-                        icon: '<i class="fas fa-times"></i>',
-                        label: "Cancel",
-                    },
-                },
-                default: "Yes",
-            }).render(true);
-        });
     }
 
     private onAddCoinsPopup(event: JQuery.ClickEvent) {
