+++ types/foundry/common/utils/collection.d.ts
@@ -65,7 +65,7 @@ declare module foundry {
              * c.get("d"); // null
              * c.get("d", {strict: true}); // throws Error
              */
-            get<T extends V = V>(key: string, { strict }: { strict: true }): T;
+            get<T extends V = V>(key: string | null | undefined, { strict }: { strict: true }): T;
             get<T extends V = V>(key: string, { strict }?: { strict?: boolean }): T | undefined;
 
             /**
